added wind direction in degrees, fixed mixup winds/windk
