+++ .homeychangelog.json
@@ -10,5 +10,9 @@
   "0.1.9": {
     "en": "Updated settings page to be a little more user friendly",
     "nl": "Settings pagina aangepast om iets gebruikersvriendelijker te zijn."
+  },
+  "0.1.11": {
+    "en": "Added wind direction in degrees",
+    "nl": "Windrichting in graden toegevoegd"
   }
 }
