+++ app.json
@@ -1,6 +1,6 @@
 {
   "id": "nl.bochove.knmi",
-  "version": "0.1.9",
+  "version": "0.1.11",
   "compatibility": ">=5.0.0",
   "sdk": 3,
   "name": {
